// src/ts-ir/types/IRTypes.ts
/**
 * Farpy - A programming language
 *
 * Copyright (c) 2025 Fernando (FernandoTheDev)
 *
 * This software is licensed under the MIT License.
 * See the LICENSE file in the project root for full license information.
 */
export interface IRValue {
  value: string;
  type: string; // Ex.: "i32", "i32*", etc.
}